# backend/tests/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentHub Multi-Agent Chatbot Framework",
    "description": "Production-ready multi-tenant chatbot framework using LangChain 0.3+",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/{tenant_id}/chat": {
      "post": {
        "tags": [
          "chat",
          "chat"
        ],
        "summary": "Chat Endpoint",
        "description": "Process user message and return agent response.\n\nFlow:\n1. Validate tenant access\n2. Create or retrieve session\n3. Save user message\n4. Route to SupervisorAgent\n5. Save assistant response\n6. Track performance metrics",
        "operationId": "chat_endpoint_api__tenant_id__chat_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Tenant UUID",
              "title": "Tenant Id"
            },
            "description": "Tenant UUID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/{tenant_id}/test/chat": {
      "post": {
        "tags": [
          "chat",
          "chat"
        ],
        "summary": "Test Chat Endpoint",
        "description": "Process user message and return agent response for testing.\nThis endpoint bypasses authentication.",
        "operationId": "test_chat_endpoint_api__tenant_id__test_chat_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Tenant UUID",
              "title": "Tenant Id"
            },
            "description": "Tenant UUID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message",
            "description": "User message content"
          },
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "User identifier (external user ID from auth system)"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "Existing session UUID as string for follow-up messages"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata",
            "description": "Additional metadata (e.g., jwt_token for external API calls)"
          }
        },
        "type": "object",
        "required": [
          "message",
          "user_id"
        ],
        "title": "ChatRequest",
        "description": "Request schema for chat endpoint."
      },
      "ChatResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "description": "Session UUID"
          },
          "message_id": {
            "type": "string",
            "title": "Message Id",
            "description": "Message UUID"
          },
          "response": {
            "additionalProperties": true,
            "type": "object",
            "title": "Response",
            "description": "Agent response data (structure varies by agent)"
          },
          "agent": {
            "type": "string",
            "title": "Agent",
            "description": "Agent that processed the request"
          },
          "intent": {
            "type": "string",
            "title": "Intent",
            "description": "Detected user intent"
          },
          "format": {
            "type": "string",
            "title": "Format",
            "description": "Output format type (text/json/table)"
          },
          "renderer_hint": {
            "additionalProperties": true,
            "type": "object",
            "title": "Renderer Hint",
            "description": "UI rendering hints"
          },
          "metadata": {
            "$ref": "#/components/schemas/ResponseMetadata",
            "description": "Response metadata with LLM model info, tool calls, and extracted entities"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "message_id",
          "response",
          "agent",
          "intent",
          "format",
          "metadata"
        ],
        "title": "ChatResponse",
        "description": "Response schema for chat endpoint."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LLMModelInfo": {
        "properties": {
          "llm_model_id": {
            "type": "string",
            "title": "Llm Model Id",
            "description": "LLM model UUID"
          },
          "model_class": {
            "type": "string",
            "title": "Model Class",
            "description": "LLM class name (e.g., ChatOpenAI, ChatOpenRouter)"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name",
            "description": "Model name (e.g., openai/gpt-4o-mini)"
          }
        },
        "type": "object",
        "required": [
          "llm_model_id",
          "model_class",
          "model_name"
        ],
        "title": "LLMModelInfo",
        "description": "LLM model information."
      },
...
      "ResponseMetadata": {
        "properties": {
          "agent_id": {
            "type": "string",
            "title": "Agent Id",
            "description": "Agent UUID that processed the request"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id",
            "description": "Tenant UUID"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms",
            "description": "Request processing duration in milliseconds"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status",
            "description": "Processing status"
          },
          "llm_model": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LLMModelInfo"
              },
              {
                "type": "null"
              }
            ],
            "description": "LLM model information"
          },
          "tool_calls": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ToolCallInfo"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Calls",
            "description": "Tools called during processing"
          },
          "extracted_entities": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extracted Entities",
            "description": "Entities extracted from user message (e.g., tax_code, salesman)"
          }
        },
        "type": "object",
        "required": [
          "agent_id",
          "tenant_id"
        ],
        "title": "ResponseMetadata",
        "description": "Extended metadata for chat responses."
      },
...
      "ToolCallInfo": {
        "properties": {
          "tool_name": {
            "type": "string",
            "title": "Tool Name",
            "description": "Name of the tool that was called"
          },
          "tool_args": {
            "additionalProperties": true,
            "type": "object",
            "title": "Tool Args",
            "description": "Arguments passed to the tool"
          },
          "tool_id": {
            "type": "string",
            "title": "Tool Id",
            "description": "Unique identifier for this tool call"
          }
        },
        "type": "object",
        "required": [
          "tool_name",
          "tool_args",
          "tool_id"
        ],
        "title": "ToolCallInfo",
        "description": "Tool call information."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}